Add the posts specs for the index and show methods. rename the posts_spec.rb file to posts_request_spec.rb.
